requirements.txt
# ML
catboost==1.2.7
lightgbm==4.6.0
scikit-learn==1.6.1
hyperopt==0.2.7
category_encoders==2.8.1

# Util
apache-airflow==3.0.1
numpy==1.26.4
pandas==2.2.3
geopandas==1.0.1
scipy==1.15.2
matplotlib==3.10.1
seaborn==0.13.2
openpyxl
xlrd
requests==2.32.3
folium==0.19.5
joblib==1.4.2
shap==0.47.1
shapely==2.0.7
wandb==0.19.9
geopy==2.4.1
xyzservices
mlxtend
python-dotenv
tqdm
xmltodict
fire

# Web
Flask==3.1.0
selenium==4.33.0
fastapi==0.115.12
streamlit==1.45.1
s3fs==2025.5.1
flask-cors
waitress